tests(integration): add integration tests (#21)
* tests(integration): add integration tests

The integration test flow starts from ci.yml, the tests are defined in
integration.test.js, where the GitHub API is used to interact with the
workflow (integration.yml) where the scheduler is used to balance and
run some dummy tests.

The test results are evaluated by checking the busted events generated
during the test runs. These are collected and dumped to file using a
busted helper. Artifacts are used to save the files generated during
execution, that the test need to check to confirm correct execution.

* review suggestions - style improvements

Co-authored-by: Hans Hübner <hans.huebner@gmail.com>